check_bacula.py: Added a check for all jobs, wrote some better errors.
If you generally try to run a nagios check and it fails, nagios won't
tell you, the OS will. Hopefully printing those error messages as a CRIT
warning will alleviate this issue.

I switched the db cursor to return a dictionary, so all new db rows now
need to be referenced by their row header (Ex. row["name"]).

There is still the issue with database connection note going to be
closed. The conn.close and cursor.close calls need to be refactored to
come before the return, but after the checks.